% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.wf_propensity_weights}
\alias{plot.wf_propensity_weights}
\title{Plot propensity overlap and covariate balance}
\usage{
\method{plot}{wf_propensity_weights}(x, lang = NULL, ...)
}
\arguments{
\item{x}{A \code{wf_propensity_weights} object.}

\item{lang}{Output language.}

\item{...}{Additional arguments passed to the initial plot calls.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot propensity overlap and covariate balance
}
