% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{print.wf_attrition_weights}
\alias{print.wf_attrition_weights}
\title{Print attrition weights}
\usage{
\method{print}{wf_attrition_weights}(x, ...)
}
\arguments{
\item{x}{A \code{wf_attrition_weights} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print attrition weights
}
