% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-trim.R
\name{print.wf_auto_trim}
\alias{print.wf_auto_trim}
\title{Print an automatic trim recommendation}
\usage{
\method{print}{wf_auto_trim}(x, ...)
}
\arguments{
\item{x}{A \code{wf_auto_trim} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an automatic trim recommendation
}
