% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoweigh.R
\name{print.wf_autoweigh_result}
\alias{print.wf_autoweigh_result}
\title{Print a guided weighting result}
\usage{
\method{print}{wf_autoweigh_result}(x, ...)
}
\arguments{
\item{x}{A \code{wf_autoweigh_result} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a guided weighting result
}
