% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{print.wf_blend_result}
\alias{print.wf_blend_result}
\title{Print a blend result.}
\usage{
\method{print}{wf_blend_result}(x, ...)
}
\arguments{
\item{x}{A \code{wf_blend_result} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a blend result.
}
