% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{print.wf_collapse_plan}
\alias{print.wf_collapse_plan}
\title{Print a collapse plan}
\usage{
\method{print}{wf_collapse_plan}(x, ...)
}
\arguments{
\item{x}{A \code{wf_collapse_plan} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a collapse plan
}
