% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-data.R
\name{print.wf_design_data}
\alias{print.wf_design_data}
\title{Print outcome-free survey design data}
\usage{
\method{print}{wf_design_data}(x, ...)
}
\arguments{
\item{x}{A \code{wf_design_data} object.}

\item{...}{Reserved for future use.}
}
\value{
\code{x}, invisibly.
}
\description{
Print outcome-free survey design data
}
