% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.wf_diagnostics}
\alias{print.wf_diagnostics}
\title{Print diagnostics}
\usage{
\method{print}{wf_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{wf_diagnostics} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print diagnostics
}
