% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{print.wf_impact}
\alias{print.wf_impact}
\title{Print post-lock outcome impact}
\usage{
\method{print}{wf_impact}(x, ...)
}
\arguments{
\item{x}{A \code{wf_impact} object.}

\item{...}{Reserved for future use.}
}
\value{
\code{x}, invisibly.
}
\description{
Print post-lock outcome impact
}
