% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{print.wf_influence}
\alias{print.wf_influence}
\title{Print influence diagnostics}
\usage{
\method{print}{wf_influence}(x, ...)
}
\arguments{
\item{x}{A \code{wf_influence} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print influence diagnostics
}
