% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest-ladder.R
\name{print.wf_ladder_draft}
\alias{print.wf_ladder_draft}
\title{Print a collapse-ladder draft}
\usage{
\method{print}{wf_ladder_draft}(x, ...)
}
\arguments{
\item{x}{A \code{wf_ladder_draft} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a collapse-ladder draft
}
