% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.wf_pipeline}
\alias{print.wf_pipeline}
\title{Print a pipeline specification}
\usage{
\method{print}{wf_pipeline}(x, ...)
}
\arguments{
\item{x}{A \code{wf_pipeline} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a pipeline specification
}
