% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat-plan.R
\name{print.wf_poststrat_plan}
\alias{print.wf_poststrat_plan}
\title{Print a post-stratification plan}
\usage{
\method{print}{wf_poststrat_plan}(x, ...)
}
\arguments{
\item{x}{A \code{wf_poststrat_plan} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a post-stratification plan
}
