% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck.R
\name{print.wf_precheck}
\alias{print.wf_precheck}
\title{Print a precheck result}
\usage{
\method{print}{wf_precheck}(x, ...)
}
\arguments{
\item{x}{A \code{wf_precheck} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a precheck result
}
