% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.wf_quality_report}
\alias{print.wf_quality_report}
\title{Print a weighting quality report}
\usage{
\method{print}{wf_quality_report}(x, ...)
}
\arguments{
\item{x}{A \code{wf_quality_report} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a weighting quality report
}
