% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{print.wf_replicate_weights}
\alias{print.wf_replicate_weights}
\title{Print replicate weights}
\usage{
\method{print}{wf_replicate_weights}(x, ...)
}
\arguments{
\item{x}{A \code{wf_replicate_weights} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print replicate weights
}
