% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{print.wf_variance_result}
\alias{print.wf_variance_result}
\title{Print a variance result}
\usage{
\method{print}{wf_variance_result}(x, ...)
}
\arguments{
\item{x}{A \code{wf_variance_result} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a variance result
}
