% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{print.wf_weights}
\alias{print.wf_weights}
\title{Print calibrated weights}
\usage{
\method{print}{wf_weights}(x, ...)
}
\arguments{
\item{x}{A \code{wf_weights} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print calibrated weights
}
