% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat-plan.R
\name{summary.wf_poststrat_plan}
\alias{summary.wf_poststrat_plan}
\title{Summarize a post-stratification plan}
\usage{
\method{summary}{wf_poststrat_plan}(object, ...)
}
\arguments{
\item{object}{A \code{wf_poststrat_plan} object.}

\item{...}{Unused.}
}
\value{
The plan diagnostics data frame.
}
\description{
Summarize a post-stratification plan
}
