% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{wf_dims}
\alias{wf_dims}
\title{Declare calibration dimensions}
\usage{
wf_dims(..., .collapse = list())
}
\arguments{
\item{...}{Named dimension-level pairs. Use \code{NULL} to infer levels from the target.}

\item{.collapse}{Named list of collapse ladders.}
}
\value{
A \code{wf_dims} object.
}
\description{
Declare calibration dimensions
}
\examples{
dims <- wf_dims(gender = c("female", "male"), age = c("young", "old"))
dims
}
